Fix bugs with the packaging operator
* add some extra error checking to make
sure the files we are copying do exist
* we need to handle library files coming
from Blender's install directory slightly
different.

+++ rman_operators/rman_operators_utils.py
@@ -84,6 +84,11 @@ class PRMAN_OT_Renderman_Package(Operator):
     @classmethod
     def poll(cls, context):
         return context.engine == "PRMAN_RENDER"
+    
+    def is_blender_install_path(self, filepath):
+        blender_binary = bpy.app.binary_path
+        bpath = os.path.normpath(os.path.dirname(blender_binary))
+        return bpy.path.is_subdir(filepath, bpath)
 
     def execute(self, context):
 
@@ -130,9 +135,23 @@ class PRMAN_OT_Renderman_Package(Operator):
                         rfb_log().debug("Cannot remove: %s" % self.filepath)
                         pass
                     return {'FINISHED'}
-                subdir = os.path.dirname(lib.filepath).replace('//', '', 1)
-                dst_path = os.path.join(self.directory, subdir)
-                shutil.copytree(os.path.dirname(real_path), dst_path)
+                if self.is_blender_install_path(real_path):
+                    # we need to do something slightly different
+                    # for library files coming from Blender's install directory
+                    subdir = os.path.join('blender_libs')
+                    lib_file = os.path.basename(real_path)
+                    libpath = os.path.join('//', '.', subdir, lib_file)
+                    lib.filepath = libpath
+                    dst_path = os.path.join(self.directory, subdir)
+                    if not os.path.exists(dst_path):
+                        os.mkdir(dst_path)
+                    if dst_path not in remove_dirs:
+                        remove_dirs.append(dst_path)
+                    shutil.copy(real_path, dst_path)
+                else:
+                    subdir = os.path.dirname(lib.filepath).replace('//', '', 1)
+                    dst_path = os.path.join(self.directory, subdir)
+                    shutil.copytree(os.path.dirname(real_path), dst_path)
 
             # get all directories and files that were copied from the libraries
             for root, dirnames, files in os.walk(self.directory):
@@ -282,13 +301,14 @@ class PRMAN_OT_Renderman_Package(Operator):
         # volumes
         for db in bpy.data.volumes:
             openvdb_file = filepath_utils.get_real_path(db.filepath)
-            bfile = os.path.basename(openvdb_file)
-            diskpath = os.path.join(assets_dir, bfile)
-            shutil.copyfile(openvdb_file, diskpath)      
-            #setattr(db, 'filepath', '//./assets/%s' % bfile)  - agentyRANCH
-            setattr(db, 'filepath', '//assets/%s' % bfile)            
-            z.write(diskpath, arcname=os.path.join('assets', bfile))               
-            remove_files.append(diskpath)
+            if os.path.exists(openvdb_file):
+                bfile = os.path.basename(openvdb_file)
+                diskpath = os.path.join(assets_dir, bfile)
+                shutil.copyfile(openvdb_file, diskpath)      
+                #setattr(db, 'filepath', '//./assets/%s' % bfile)  - agentyRANCH
+                setattr(db, 'filepath', '//assets/%s' % bfile)            
+                z.write(diskpath, arcname=os.path.join('assets', bfile))               
+                remove_files.append(diskpath)
             
         # Caches #  - agentyRANCH
         # https://docs.blender.org/manual/fr/dev/animation/constraints/transform/transform_cache.html
@@ -296,12 +316,13 @@ class PRMAN_OT_Renderman_Package(Operator):
             # Change get_real_path with filesystem_path  - agentyRANCH
             # (resolve blender relative path for shutil)
             cache_file = filepath_utils.filesystem_path(cache.filepath)
-            bfile = os.path.basename(cache_file)
-            diskpath = os.path.join(assets_dir, bfile)
-            shutil.copyfile(cache_file, diskpath)      
-            setattr(cache, 'filepath', '//assets/%s' % bfile)            
-            z.write(diskpath, arcname=os.path.join('assets', bfile))               
-            remove_files.append(diskpath)            
+            if os.path.exists(cache_file):
+                bfile = os.path.basename(cache_file)
+                diskpath = os.path.join(assets_dir, bfile)
+                shutil.copyfile(cache_file, diskpath)      
+                setattr(cache, 'filepath', '//assets/%s' % bfile)            
+                z.write(diskpath, arcname=os.path.join('assets', bfile))               
+                remove_files.append(diskpath)            
 
         # archives etc.
         for ob in bpy.data.objects:
@@ -309,30 +330,33 @@ class PRMAN_OT_Renderman_Package(Operator):
             if rman_type == 'DELAYED_LOAD_ARCHIVE':
                 rm = ob.renderman
                 rib_path = string_utils.expand_string(rm.path_archive)
-                bfile = os.path.basename(rib_path)
-                diskpath = os.path.join(assets_dir, bfile)
-                shutil.copyfile(rib_path, diskpath)  
-                setattr(rm, 'path_archive', os.path.join('<blend_dir>', 'assets', bfile))
-                z.write(diskpath, arcname=os.path.join('assets', bfile))
-                remove_files.append(diskpath)    
+                if os.path.exists(rib_path):
+                    bfile = os.path.basename(rib_path)
+                    diskpath = os.path.join(assets_dir, bfile)
+                    shutil.copyfile(rib_path, diskpath)  
+                    setattr(rm, 'path_archive', os.path.join('<blend_dir>', 'assets', bfile))
+                    z.write(diskpath, arcname=os.path.join('assets', bfile))
+                    remove_files.append(diskpath)    
             elif rman_type == 'ALEMBIC':
                 rm = ob.renderman
                 abc_filepath = string_utils.expand_string(rm.abc_filepath)
-                bfile = os.path.basename(abc_filepath)
-                diskpath = os.path.join(assets_dir, bfile)
-                shutil.copyfile(abc_filepath, diskpath)  
-                setattr(rm, 'abc_filepath', os.path.join('<blend_dir>', 'assets', bfile))
-                z.write(diskpath, arcname=os.path.join('assets', bfile))
-                remove_files.append(diskpath)   
+                if os.path.exists(abc_filepath):
+                    bfile = os.path.basename(abc_filepath)
+                    diskpath = os.path.join(assets_dir, bfile)
+                    shutil.copyfile(abc_filepath, diskpath)  
+                    setattr(rm, 'abc_filepath', os.path.join('<blend_dir>', 'assets', bfile))
+                    z.write(diskpath, arcname=os.path.join('assets', bfile))
+                    remove_files.append(diskpath)   
             elif rman_type == 'BRICKMAP':  
                 rm = ob.renderman
                 bkm_filepath = string_utils.expand_string(rm.bkm_filepath)
-                bfile = os.path.basename(bkm_filepath)
-                diskpath = os.path.join(assets_dir, bfile)
-                shutil.copyfile(bkm_filepath, diskpath)  
-                setattr(rm, 'bkm_filepath', os.path.join('<blend_dir>', 'assets', bfile))
-                z.write(diskpath, arcname=os.path.join('assets', bfile))
-                remove_files.append(diskpath)                                
+                if os.path.exists(bkm_filepath):
+                    bfile = os.path.basename(bkm_filepath)
+                    diskpath = os.path.join(assets_dir, bfile)
+                    shutil.copyfile(bkm_filepath, diskpath)  
+                    setattr(rm, 'bkm_filepath', os.path.join('<blend_dir>', 'assets', bfile))
+                    z.write(diskpath, arcname=os.path.join('assets', bfile))
+                    remove_files.append(diskpath)                                
 
         # include disgust trace
         if self.properties.include_disgust:
@@ -354,7 +378,7 @@ class PRMAN_OT_Renderman_Package(Operator):
             context.scene.renderman.rfb_disgust = False
 
             bpy.ops.wm.save_as_mainfile(filepath=bl_filepath, copy=True, compress=False, relative_remap=False)
-            remove_files.append(bl_filepath)
+            # remove_files.append(bl_filepath)
 
             z.write(bl_filepath, arcname=bl_filename)
             context.scene.renderman.rfb_disgust = before
@@ -370,11 +394,16 @@ class PRMAN_OT_Renderman_Package(Operator):
 
         for d in remove_dirs:
             try:
+                print("REMOVING DIR: %s" % d)
                 os.removedirs(d)
             except:
                 continue
 
         bpy.ops.wm.revert_mainfile()
+        try:
+            os.remove(bl_filepath)
+        except:
+            pass
 
         return {'FINISHED'}
